Rename game-objects to objects
